requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
grpcio==1.50.0
grpcio-tools==1.50.0
idna==3.4
protobuf==4.21.9
pyTelegramBotAPI==4.7.1
requests==2.28.1
six==1.16.0
telebot==0.0.4
urllib3==1.26.12
